Cleanup: Finally moved help-channel-messages to inside src
