+++ src/interaction-handling/select-menu-handler.ts
@@ -5,9 +5,9 @@ import { SelectMenuHandlerProps } from './handler-interface.js';
 import { ExpectedParseErrors } from './interaction-constants/expected-interaction-errors.js';
 import { SelectMenuNames } from './interaction-constants/interaction-names.js';
 import { AttendingServer } from '../attending-server/base-attending-server.js';
-import { adminCommandHelpMessages } from '../../help-channel-messages/AdminCommands.js';
-import { helperCommandHelpMessages } from '../../help-channel-messages/HelperCommands.js';
-import { studentCommandHelpMessages } from '../../help-channel-messages/StudentCommands.js';
+import { adminCommandHelpMessages } from '../help-channel-messages/AdminCommands.js';
+import { helperCommandHelpMessages } from '../help-channel-messages/HelperCommands.js';
+import { studentCommandHelpMessages } from '../help-channel-messages/StudentCommands.js';
 import { ReturnToHelpMainAndSubMenuButton } from './shared-interaction-functions.js';
 import { AccessLevelRole } from '../models/access-level-roles.js';
 import { QuickStartLoggingChannel } from '../attending-server/quick-start-pages.js';
